pyproject.toml
[project]
name = "api-netescola"
version = "0.1.0"
description = ""
authors = [
    {name = "4trx0",email = "rickmartim41@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "beanie (>=2.0.0,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
